+++ app/src/main/java/com/playgilround/schedule/client/singin/SignInActivity.java
@@ -1,13 +1,18 @@
 package com.playgilround.schedule.client.singin;
 
+import android.Manifest;
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Button;
 
+import com.gun0912.tedpermission.PermissionListener;
+import com.gun0912.tedpermission.TedPermission;
 import com.playgilround.schedule.client.R;
 import com.playgilround.schedule.client.activity.MainActivity;
 import com.playgilround.schedule.client.activity.ScheduleCalendarActivity;
 import com.playgilround.schedule.client.model.User;
+
+import java.util.List;
 
 import androidx.appcompat.app.AppCompatActivity;
 import butterknife.BindView;
@@ -18,18 +23,35 @@ public class SignInActivity extends AppCompatActivity implements SignInContract.
 
     private SignInContract.Presenter mPresenter;
 
+    private String[] PERMISSION_STORAGE = {
+            Manifest.permission.ACCESS_FINE_LOCATION,
+            Manifest.permission.ACCESS_COARSE_LOCATION,
+            Manifest.permission.INTERNET
+    };
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_login);
+
+        TedPermission.with(this)
+                .setPermissionListener(new PermissionListener() {
+                    @Override
+                    public void onPermissionGranted() {
+
+                    }
+
+                    @Override
+                    public void onPermissionDenied(List<String> deniedPermissions) {
+
+                    }
+                })
+                .setDeniedMessage(getString(R.string.message_check_permission))
+                .setPermissions(PERMISSION_STORAGE)
+                .check();
+
         ButterKnife.bind(this);
 
         new SignInPresenter(this);
-    }
-
-    @OnClick(R.id.btn_login)
-    public void signIn() {
-        startActivity(new Intent(this, ScheduleCalendarActivity.class));
     }
 
     @Override
